+++ src/main/scala/name.lakhin.eliah.projects/papacarlo/syntax/Cache.scala
@@ -56,6 +56,7 @@ final class Cache(syntax: Syntax,
                 new Cache(syntax, descendantFragment, descendant)
 
       syntax.onNodeMerge.trigger(node)
+      node.update()
     }
   }